Added a friend count to free and freinds. Minor bug.
This little change took a disturbingly long amount of time to complete! It was mostly due to a bug that would keep calling the fetchData function infinitely with no obvious disturbance on the user's end. On second thought, I could've left it in there but that wouldn't be ethical to leave a potential memory leak like that. 

For now, instead of infinite, fetching and retrieving accessToken is being called 4 times. I know it's adding 4 times the demand on the database, but I will have fix it later.